+++ Sources/SentryObjC/Public/SentryObjCUserFeedbackConfiguration.h
@@ -33,6 +33,7 @@ NS_ASSUME_NONNULL_BEGIN
  */
 @property (nonatomic) BOOL showFormForScreenshots;
 
+#    if !SDK_V10
 /**
  * Install a hook for the specified button to show the form when it is pressed.
  * @note Default value is @c nil.
@@ -43,6 +44,7 @@ NS_ASSUME_NONNULL_BEGIN
     __attribute__((deprecated("The custom User Feedback button configuration is deprecated and "
                               "will be removed in v10. Add your own button action and call "
                               "[[SentryObjCSDK feedback] show] instead.")));
+#    endif
 
 /**
  * Configuration settings specific to the managed UI form to gather user input.